Split multi-type files, and nest what that unblocks
Four files declaring more than one top-level type are split into
one-type-per-file, which is what Consequent's L2 asks for anyway, and
the nesting they were blocking follows: `Http2.Connection`,
`Http2.Stream`, `Mcp.Server`, `Mcp.Session`, `Zip.Error`, `Zip.Event`
and `Zip.Handle`.

Two of the four splits do not survive, and the second is the more
interesting.

`Zip.Openable` and `Zip.DataOpenable` shadow the `Openable` they
extend, so the nested types would have to qualify their own base.
These are single-given backing types anyway, which the proposal's R7
would rather see inlined than nested, so they stay top-level pending
that.

Hpack is the surprise: hoisting `HpackEntry` out of `HpackTable.scala`
broke capture inference on its own, before any nesting —
`HpackTable.static` infers `Array[HpackEntry]^{}` where the use site
then demands `Array[HpackEntry^'s1]^{any}`, and `Array` is invariant.
The two types were sharing a file for a reason, even though nothing
recorded it. `TarHeader` failed the same way when nested; this shows
the sensitivity is to the *file boundary*, not only to the enclosing
object, so a hoist is not the free preparatory step it appears to be.
Hpack is restored to its original single file.
