+++ lib/anthology/src/linker/anthology.Bundler.scala
@@ -76,7 +76,7 @@ object Bundler:
 
   private[anthology] def assemble
     ( classpath: LocalClasspath, jarfile: Path on Linux, main: Optional[Fqcn] )
-  :   Path on Linux raises ZipError raises Path.Error raises IoError raises StreamError =
+  :   Path on Linux raises Zip.Error raises Path.Error raises IoError raises StreamError =
 
     val manifest =
       main.let(MainClass(_)).let: main =>
