--- lib/synesthesia/src/core/synesthesia.McpServer.scala
@@ -1,89 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.64.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package synesthesia
-
-import scala.annotation.*
-import scala.collection.mutable as scm
-
-import anticipation.*
-import contingency.*
-import distillate.*
-import fulminate.*
-import gossamer.*
-import inimitable.*
-import jacinta.*
-import obligatory.*
-import parasite.*
-import prepositional.*
-import revolution.*
-import rudiments.*
-import scintillate.*
-import telekinesis.*
-import urticose.*
-import vacuous.*
-import wisteria.*
-
-trait McpServer():
-  import Mcp.*
-
-  private val sessions: scm.HashMap[Text, Session] = scm.HashMap()
-
-  type Session <: McpSession
-  type Origin = Mcp.Client
-
-  def session(id: Text): Session = sessions.establish(id)(initialize())
-  def initialize(): Session
-
-  private given mcpSessionId: ("mcpSessionId" is Directive of Text) = identity(_)
-
-
-  def serve(using this.type is Mcp.Specification, Monitor, Probate, Online, Http.Request)
-  :   Http.Response =
-
-    recover:
-      case error @ JsonRpcError(_, _, _) =>
-        import hieroglyph.charEncoders.utf8Encoder
-        Http.Response(Unfulfilled(t"JSON-RPC error: ${error.message.text}"))
-
-    . protect:
-        val sessionId = request.headers.mcpSessionId.prim.or(Uuid().encode)
-        val interface: Mcp.Interface = Mcp.Interface(sessionId, this)
-        Mcp.send(sessionId, this, interface)(JsonRpc.serve(interface))
-
-
-  def name: Text
-  def description: Text
-  def version: Semver
-  def prompts: List[Prompt]
-
-trait McpSession
