+++ lib/telekinesis/src/http2/cordillera.Hpack.scala
@@ -42,6 +42,7 @@ import rudiments.*
 import vacuous.*
 
 import Http2.Error.Reason
+import scala.caps
 
 // HPACK header-block compression and decompression (RFC 7541). A `Hpack` instance
 // owns one direction's dynamic table; a connection keeps one for decoding inbound
