+++ lib/telekinesis/src/http2/cordillera.Http2ServerConnection.scala
@@ -138,7 +138,7 @@ object Http2ServerConnection:
               conn.streams.remove(id)
 
           case None =>
-            val stream = Http2Stream(id)
+            val stream = Http2.Stream(id)
             conn.streams(id) = stream
             stream.acceptHeaders(decoder.decode(block))
             if endStream then stream.end() else ()
@@ -196,7 +196,7 @@ class Http2ServerConnection(duplex: Duplex^)(using Monitor, Probate):
   // the reader/writer daemons (and reaches `close`) as a neutral `AnyRef` rim.
   private val duplexRef: AnyRef = duplex.asInstanceOf[AnyRef]
 
-  private[cordillera] val streams: scc.TrieMap[Int, Http2Stream] = scc.TrieMap()
+  private[cordillera] val streams: scc.TrieMap[Int, Http2.Stream] = scc.TrieMap()
   private val outbound: Relay[Frame] = Relay()
   private[cordillera] val started: Promise[Unit] = Promise()
 
@@ -209,7 +209,7 @@ class Http2ServerConnection(duplex: Duplex^)(using Monitor, Probate):
 
   // Streams opened by the client, in arrival order; the serve loop takes each
   // and runs its handler. Stopped when the connection ends.
-  private[cordillera] val accepted: Relay[Http2Stream] = Relay()
+  private[cordillera] val accepted: Relay[Http2.Stream] = Relay()
 
   private[cordillera] def send(frame: Frame): Unit = outbound.put(frame)
 
@@ -223,7 +223,7 @@ class Http2ServerConnection(duplex: Duplex^)(using Monitor, Probate):
     accepted.stop()
 
   private val (writer, reader) =
-    // As `Http2Connection`: no aliased writer between containment and body.
+    // As `Http2.Connection`: no aliased writer between containment and body.
     scala.caps.unsafe.unsafeAssumeSeparate:
      contain:
       case _ => tearDown(); Remedy.Accept
@@ -253,7 +253,7 @@ class Http2ServerConnection(duplex: Duplex^)(using Monitor, Probate):
 
           // The server's first read: consume and validate the client
           // connection preface before frame-parsing.
-          frameReader.expectPreface(Http2Connection.connectionPreface)
+          frameReader.expectPreface(Http2.Connection.connectionPreface)
 
           val decoder = Hpack()
           var continue = true
@@ -278,7 +278,7 @@ class Http2ServerConnection(duplex: Duplex^)(using Monitor, Probate):
   // reader-driven serve loop; returns when the connection ends. The handler
   // typically spawns a per-stream fiber so requests multiplexed on the one
   // connection are served concurrently.
-  def eachStream(handler: Http2Stream => Unit): Unit =
+  def eachStream(handler: Http2.Stream => Unit): Unit =
     accepted.stream.records.each(handler)
 
   // Send a response header block on `streamId`, encoded with a fresh
