+++ lib/telekinesis/src/http2/telekinesis_http2.scala
@@ -107,7 +107,7 @@ extends Sessional:
               scala.caps.unsafe.unsafeAssumeSeparate:
                unsafely:
                 supervise:
-                  val connection = Http2Connection(duplex)
+                  val connection = Http2.Connection(duplex)
 
                   try
                     connection.start()
