+++ lib/zeppelin/src/core/zeppelin.ZipOpenable.scala
@@ -32,73 +32,47 @@
                                                                                                   */
 package zeppelin
 
-import scala.caps
-
+import java.io as ji
+import java.util.zip as juz
 import proscenium.compat.*
-
-import java.nio as jn
-import java.nio.channels as jnc
-import java.nio.file as jnf
-
 import anticipation.*
-import aperture.*
 import contingency.*
+import galilei.*
+import gossamer.*
+import nomenclature.*
 import prepositional.*
 import rudiments.*
 import serpentine.*
-import vacuous.*
+import pneumatic.*
+import turbulence.*
 import zephyrine.*
+import vacuous.*
+import fulminate.*
+import scala.caps
+import java.nio as jn
+import java.nio.channels as jnc
+import java.nio.file as jnf
+import aperture.*
 
-// The scoped capability provided by opening an archive as `Zip`: `path.open[Zip]()`. Unlike a
-// detached `Zipfile` (whose `FileSource` re-opens the file for every read), a `ZipHandle` reads
-// through a single channel held open for the duration of the block, so entry payloads resolve
-// with no per-read open/close cost — and, correspondingly, must be consumed within the scope.
-// (Zeppelin is not yet capture-checked, so the confinement is enforced only for callers
-// compiled with capture checking; the annotations sharpen when the module joins the rollout.)
-class ZipHandle private[zeppelin] (private[zeppelin] val zipfile: Zipfile)
-extends caps.ExclusiveCapability:
-  def entries: List[Zip.Entry] = zipfile.entries
-  def entry(ref: Path on Zip): Zip.Entry raises ZipError = zipfile.entry(ref)
-  def comment: Optional[Text] = zipfile.comment
-
-// A named class rather than an anonymous given instance, for the reasons documented on
-// galilei's `FileOpenable`. Archives open read-only: a `Write` mode is refused with
-// `ZipError.Reason.WriteUnsupported` until writing lands.
-class ZipOpenable[path: Abstractable across Paths to Text](using Tactic[ZipError])
+class ZipOpenable[path: Abstractable across Paths to Text](using Tactic[Zip.Error])
 extends Openable:
 
   type Self = path
   type Form = Zip
   type Operand = Nothing
-  type Result = ZipHandle
+  type Result = Zip.Handle
 
   def open[grants <: Grant, result]
     ( value: path, mode: Mode granting grants, flags: List[Nothing] )
-    ( block: (ZipHandle & Granting[grants]) ?=> result )
+    ( block: (Zip.Handle & Granting[grants]) ?=> result )
   :   result =
 
-    if mode.atoms.has(Write) then abort(ZipError(ZipError.Reason.WriteUnsupported))
+    if mode.atoms.has(Write) then abort(Zip.Error(Zip.Error.Reason.WriteUnsupported))
 
     val channel =
       jnc.FileChannel.open(jnf.Path.of(value.generic.s), jnf.StandardOpenOption.READ).nn
 
     try
       val zipfile = Zipfile.parse(Zipfile.ChannelSource(channel))
-      block(using new ZipHandle(zipfile) with Granting[grants] {})
+      block(using new Zip.Handle(zipfile) with Granting[grants] {})
     finally channel.close()
-
-// Opens an in-memory archive; no channel is involved, but access is scoped all the same, for
-// consistency with every other form of `Zip` target.
-class ZipDataOpenable(using Tactic[ZipError]) extends Openable:
-  type Self = Data
-  type Form = Zip
-  type Operand = Nothing
-  type Result = ZipHandle
-
-  def open[grants <: Grant, result]
-    ( value: Data, mode: Mode granting grants, flags: List[Nothing] )
-    ( block: (ZipHandle & Granting[grants]) ?=> result )
-  :   result =
-
-    if mode.atoms.has(Write) then abort(ZipError(ZipError.Reason.WriteUnsupported))
-    block(using new ZipHandle(Zipfile.parse(Zipfile.DataSource(value))) with Granting[grants] {})
