+++ lib/zeppelin/src/core/zeppelin_core.scala
@@ -34,4 +34,4 @@ package zeppelin
 
 // The contextual archive handle within an `open[Zip]` block, in the manner of galilei's `file`.
 // Transparent inline so the handle's precise (grant-refined, capturing) type is preserved.
-transparent inline def zip(using handle: ZipHandle): handle.type = handle
+transparent inline def zip(using handle: Zip.Handle): handle.type = handle
